Online-LIbrary
The E-Library project is a web-based application developed using Java Servlets, HTML, and CSS. It
provides users with an online platform to access and manage a digital library of books, journals, and
other educational resources. The project involves designing an interactive user interface using HTML
and CSS and implementing server-side functionalities using Java Servlets for user authentication,
database connectivity, and book management. The application aims to provide users with a convenient
and user friendly digital library experience, making it easy for them to access and manage educational
materials online.